// src/redux/actions/types.ts
export enum ActionTypes {
  GET_ALL_PUBLICATIONS = "GET_ALL_PUBLICATIONS",
  GET_ALL_PUBLICATIONS_LOADING = "GET_ALL_PUBLICATIONS_LOADING",
  GET_ALL_PUBLICATIONS_ERROR = "GET_ALL_PUBLICATIONS_ERROR",

  GET_ALL_CATEGORIES = "GET_ALL_CATEGORIES",
  GET_ALL_CATEGORIES_LOADING = "GET_ALL_CATEGORIES_LOADING",
  GET_ALL_CATEGORIES_ERROR = "GET_ALL_CATEGORIES_ERROR",

  GET_ONE_PUBLICATIONS = "GET_ONE_PUBLICATIONS",
  GET_ONE_PUBLICATIONS_LOADING = "GET_ONE_PUBLICATIONS_LOADING",
  GET_ONE_PUBLICATIONS_ERROR = "GET_ONE_PUBLICATIONS_ERROR",
}